Move old `mangled` methods back to the McM class
The current behavior changed the access signature to the `BaseClient`
class. Move this methods back again so that cmpatibility is kept the
same as the old client version.
